+++ src/Context.js
@@ -7,11 +7,13 @@ const FormRegisterContext = React.createContext({
   setTouched: () => {},
   setInitial: () => {},
   setError: () => {},
-  update: () => {}
+  update: () => {}, 
 });
 
 const FormStateContext = React.createContext({});
-const FormApiContext = React.createContext({});
+const FormApiContext = React.createContext({
+  getFullField: () => {}
+});
 const GroupContext = React.createContext();
 const SelectContext = React.createContext();