+++ src/hooks/useField.js
@@ -31,6 +31,7 @@ function useField(field, fieldProps = {}) {
     validateOnChange,
     validateOnBlur,
     validateOnMount,
+    maskOnBlur,
     onValueChange,
     notify,
     keepState, 
@@ -45,6 +46,7 @@ function useField(field, fieldProps = {}) {
   const [touched, setTouch] = useState();
   const [initValue, setInitValue] = useState();
   const [cursor, setCursor, getCursor] = useStateWithGetter(0);
+  const [maskedValue, setMaskedValue ] = useState(value);
 
   // Grab the form register context
   const updater = useContext(FormRegisterContext);
@@ -63,6 +65,8 @@ function useField(field, fieldProps = {}) {
   // Define set value
   const setValue = (val, e) => {
     logger(`Setting ${field} to ${val}`);
+    // Initialize maked value
+    let maskedVal = val;
     // Set value to undefined if its an empty string
     if( val === '' ){
       val = undefined;
@@ -72,12 +76,14 @@ function useField(field, fieldProps = {}) {
       val = +val;
     }
     // Call mask if it was passed
-    if(mask){
+    if(mask && !maskOnBlur){
+      maskedVal = mask(val);
       val = mask(val);
     }
     // Call format and parse if they were passed
     if(format && parse){
-      val = format(parse(val));
+      val = parse(val);
+      maskedVal = format(val);
     }
     // We only need to call validate if the user gave us one
     // and they want us to validate on change
@@ -89,8 +95,10 @@ function useField(field, fieldProps = {}) {
     if(e && e.target && e.target.selectionStart ){
       setCursor(e.target.selectionStart);
     }
+
     // Now we update the value
     setVal(val);
+    setMaskedValue(maskedVal);
     // If the user passed in onValueChange then call it!
     if( onValueChange ){
       onValueChange(val);
@@ -106,7 +114,20 @@ function useField(field, fieldProps = {}) {
     // and they want us to validate on blur
     if (validate && validateOnBlur) {
       logger(`Validating after blur ${field} ${getVal()}`);
-      setError(validate(getVal()));
+      setError(validate(getVal(), formApi.getValues()));
+    }
+    // Call mask if it was passed
+    if(mask && maskOnBlur){
+      const maskedVal = mask( getVal() );
+      // Now we update the value
+      setVal(maskedVal);
+      setMaskedValue(maskedVal);
+      // If the user passed in onValueChange then call it!
+      if( onValueChange ){
+        onValueChange(maskedVal);
+      }    
+      // Call the updater
+      updater.setValue(field, maskedVal);
     }
     setTouch(val);
     updater.setTouched(field, val);
@@ -153,20 +174,22 @@ function useField(field, fieldProps = {}) {
     error,
     touched,
     initial: value === (initialValue || initValue || ''),
+    maskedValue
   };
 
-  logger('Render', field, fieldState);
+  logger('Render', formApi.getFullField(field), fieldState);
 
   const ref = useRef(null);
 
   // We want to register and deregister this field when field name changes
   useLayoutEffect(
     () => {
-      logger('Register', field);
-      updater.register(field, fieldState, { field, fieldApi, fieldState, notify, keepState });
+      const fullField = formApi.getFullField(field);
+      logger('Register', fullField);
+      updater.register(field, fieldState, { field: fullField, fieldApi, fieldState, notify, keepState });
 
       return () => {
-        logger('Deregister', field);
+        logger('Deregister', fullField);
         updater.deregister(field);
       };
     },
@@ -177,8 +200,9 @@ function useField(field, fieldProps = {}) {
   // We want to let the controller know of changes on this field when specific props change
   useLayoutEffect(
     () => {
+      const fullField = formApi.getFullField(field);
       logger('Update', field);
-      updater.update(field, { field, fieldApi, fieldState, notify, keepState });
+      updater.update(field, { field: fullField, fieldApi, fieldState, notify, keepState });
     },
     // This is VERYYYY!! Important!
     [validate, validateOnChange, validateOnBlur, onValueChange]